+++ pyemma/coordinates/tests/test_pca.py
@@ -33,7 +33,7 @@ import numpy as np
 from pyemma.coordinates import pca, source
 from logging import getLogger
 import pyemma.util.types as types
-from six.moves import range
+
 
 
 logger = getLogger('pyemma.'+'TestPCA')