+++ pyemma/msm/estimators/maximum_likelihood_msm.py
@@ -22,8 +22,8 @@ import numpy as _np
 import warnings
 from msmtools import estimation as msmest
 
-from pyemma.util.annotators import alias, aliased, fix_docs, deprecated
-from pyemma.util.exceptions import PyEMMA_DeprecationWarning
+
+from pyemma.util.annotators import alias, aliased, fix_docs
 from pyemma.util.types import ensure_dtraj_list
 from pyemma._base.estimator import Estimator as _Estimator
 from pyemma.msm.estimators._dtraj_stats import DiscreteTrajectoryStats as _DiscreteTrajectoryStats
@@ -40,7 +40,8 @@ class _MSMEstimator(_Estimator, _MSM):
     r"""Base class for different MSM estimators given discrete trajectory statistics"""
 
     def __init__(self, lag=1, reversible=True, count_mode='sliding', sparse=False,
-                 connectivity='largest', dt_traj='1 step', score_method='VAMP2', score_k=10):
+                 connectivity='largest', dt_traj='1 step', score_method='VAMP2', score_k=10,
+                 mincount_connectivity='1/n'):
         r"""Maximum likelihood estimator for MSMs given discrete trajectory statistics
 
         Parameters
@@ -122,6 +123,12 @@ class _MSMEstimator(_Estimator, _MSM):
             The maximum number of eigenvalues or singular values used in the
             score. If set to None, all available eigenvalues will be used.
 
+        mincount_connectivity : float or '1/n'
+            minimum number of counts to consider a connection between two states.
+            Counts lower than that will count zero in the connectivity check and
+            may thus separate the resulting transition matrix. The default
+            evaluates to 1/nstates.
+
         """
         self.lag = lag
 
@@ -156,6 +163,9 @@ class _MSMEstimator(_Estimator, _MSM):
         self.score_method = score_method
         self.score_k = score_k
 
+        # connectivity
+        self.mincount_connectivity = mincount_connectivity
+
     ################################################################################
     # Generic functions
     ################################################################################
@@ -165,8 +175,7 @@ class _MSMEstimator(_Estimator, _MSM):
 
         Parameters
         ----------
-        dtrajs : list containing ndarrays(dtype=int) or ndarray(n, dtype=int)
-            or :class:`pyemma.msm.util.dtraj_states.DiscreteTrajectoryStats`
+        dtrajs : list containing ndarrays(dtype=int) or ndarray(n, dtype=int) or :class:`DiscreteTrajectoryStats <pyemma.msm.estimators._dtraj_stats.DiscreteTrajectoryStats>`
             discrete trajectories, stored as integer ndarrays (arbitrary size)
             or a single ndarray for only one trajectory.
 
@@ -184,7 +193,8 @@ class _MSMEstimator(_Estimator, _MSM):
                                     'Consider using sparse=True.')
 
         # count lagged
-        dtrajstats.count_lagged(self.lag, count_mode=self.count_mode)
+        dtrajstats.count_lagged(self.lag, count_mode=self.count_mode,
+                                mincount_connectivity=self.mincount_connectivity)
 
         # for other statistics
         return dtrajstats
@@ -353,11 +363,13 @@ class _MSMEstimator(_Estimator, _MSM):
             raise ValueError('score_cv currently only supports count modes "sliding" and "sample"')
         sliding = self.count_mode == 'sliding'
         scores = []
+        from pyemma._ext.sklearn.base import clone
+        estimator = clone(self)
         for i in range(n):
             dtrajs_split = self._blocksplit_dtrajs(dtrajs, sliding)
             dtrajs_train, dtrajs_test = cvsplit_dtrajs(dtrajs_split)
-            self.fit(dtrajs_train)
-            s = self.score(dtrajs_test, score_method=score_method, score_k=score_k)
+            estimator.fit(dtrajs_train)
+            s = estimator.score(dtrajs_test, score_method=score_method, score_k=score_k)
             scores.append(s)
         return _np.array(scores)
 
@@ -421,7 +433,6 @@ class _MSMEstimator(_Estimator, _MSM):
 
     @property
     @alias('dtrajs_full')
-    @deprecated("This attribute will be removed in the future!")
     def discrete_trajectories_full(self):
         """
         A list of integer arrays with the original (unmapped) discrete trajectories:
@@ -432,7 +443,6 @@ class _MSMEstimator(_Estimator, _MSM):
 
     @property
     @alias('dtrajs_active')
-    @deprecated("This attribute will be removed in the future!")
     def discrete_trajectories_active(self):
         """
         A list of integer arrays with the discrete trajectories mapped to the connectivity mode used.
@@ -441,39 +451,11 @@ class _MSMEstimator(_Estimator, _MSM):
         """
         self._check_is_estimated()
         # compute connected dtrajs
-        dtrajs_active = [self._full2active[dtraj] for dtraj in self._dtrajs_full]
-        return dtrajs_active
+        self._dtrajs_active = []
+        for dtraj in self._dtrajs_full:
+            self._dtrajs_active.append(self._full2active[dtraj])
 
-    @alias('compute_dtrajs_active')
-    def compute_discrete_trajectories_active(self, dtrajs):
-        """
-        A list of integer arrays with the discrete trajectories mapped to the connectivity mode used.
-        For example, for connectivity='largest', the indexes will be given within the connected set.
-        Frames that are not in the connected set will be -1.
-
-        Parameters
-        ----------
-        dtrajs : list containing ndarrays(dtype=int) or ndarray(n, dtype=int)
-            discrete trajectories, stored as integer ndarrays (arbitrary size)
-            or a single ndarray for only one trajectory.
-
-        """
-        self._check_is_estimated()
-        dtrajs = ensure_dtraj_list(dtrajs)
-        # compute connected dtrajs
-        dtrajs_active = [self._full2active[dtraj] for dtraj in dtrajs]
-        return dtrajs_active
-
-    def _check_dtrajs_arg(self, dtrajs):
-        # ensure dtrajs is given, otherwise warn the user.
-        if dtrajs is None:
-            import inspect
-            s = inspect.stack(context=3)
-            method = s[1][3]
-            warnings.warn("The dtrajs argument will be mandatory in the future for method {method}."
-                          .format(method=method), stacklevel=3, category=PyEMMA_DeprecationWarning)
-            dtrajs = self._dtrajs_full
-        return dtrajs
+        return self._dtrajs_active
 
     @property
     def count_matrix_active(self):
@@ -523,7 +505,6 @@ class _MSMEstimator(_Estimator, _MSM):
         return float(self._nstates) / float(self._nstates_full)
 
     @property
-    @deprecated('use compute_active_count_fraction')
     def active_count_fraction(self):
         """The fraction of counts in the largest connected set.
 
@@ -532,17 +513,6 @@ class _MSMEstimator(_Estimator, _MSM):
         from pyemma.util.discrete_trajectories import count_states
 
         hist = count_states(self._dtrajs_full)
-        hist_active = hist[self.active_set]
-        return float(_np.sum(hist_active)) / float(_np.sum(hist))
-
-    def compute_active_count_fraction(self, dtrajs):
-        """The fraction of counts in the largest connected set.
-        """
-        self._check_is_estimated()
-        dtrajs = ensure_dtraj_list(dtrajs)
-        from pyemma.util.discrete_trajectories import count_states
-
-        hist = count_states(dtrajs)
         hist_active = hist[self.active_set]
         return float(_np.sum(hist_active)) / float(_np.sum(hist))
 
@@ -676,7 +646,7 @@ class _MSMEstimator(_Estimator, _MSM):
     ################################################################################
     # For general statistics
     ################################################################################
-    def trajectory_weights(self, dtrajs=None):
+    def trajectory_weights(self):
         r"""Uses the MSM to assign a probability weight to each trajectory frame.
 
         This is a powerful function for the calculation of arbitrary observables in the trajectories one has
@@ -723,18 +693,16 @@ class _MSMEstimator(_Estimator, _MSM):
 
         """
         self._check_is_estimated()
-        dtrajs = self._check_dtrajs_arg(dtrajs)
-
         # compute stationary distribution, expanded to full set
         statdist_full = _np.zeros([self._nstates_full])
         statdist_full[self.active_set] = self.stationary_distribution
         # histogram observed states
         import msmtools.dtraj as msmtraj
-        hist = 1.0 * msmtraj.count_states(dtrajs)
+        hist = 1.0 * msmtraj.count_states(self.discrete_trajectories_full)
         # simply read off stationary distribution and accumulate total weight
         W = []
         wtot = 0.0
-        for dtraj in dtrajs:
+        for dtraj in self.discrete_trajectories_full:
             w = statdist_full[dtraj] / hist[dtraj]
             W.append(w)
             wtot += _np.sum(w)
@@ -748,17 +716,13 @@ class _MSMEstimator(_Estimator, _MSM):
     # HMM-based coarse graining
     ################################################################################
 
-    def hmm(self, nhidden, dtrajs=None):
+    def hmm(self, nhidden):
         """Estimates a hidden Markov state model as described in [1]_
 
         Parameters
         ----------
         nhidden : int
             number of hidden (metastable) states
-
-        dtrajs : list containing ndarrays(dtype=int) or ndarray(n, dtype=int)
-            discrete trajectories, stored as integer ndarrays (arbitrary size)
-            or a single ndarray for only one trajectory.
 
         Returns
         -------
@@ -772,7 +736,6 @@ class _MSMEstimator(_Estimator, _MSM):
 
         """
         self._check_is_estimated()
-        dtrajs = self._check_dtrajs_arg(dtrajs)
         # check if the time-scale separation is OK
         # if hmm.nstates = msm.nstates there is no problem. Otherwise, check spectral gap
         if self.nstates > nhidden:
@@ -792,7 +755,7 @@ class _MSMEstimator(_Estimator, _MSM):
         from pyemma.msm.estimators.maximum_likelihood_hmsm import MaximumLikelihoodHMSM
         estimator = MaximumLikelihoodHMSM(lag=self.lagtime, nstates=nhidden, msm_init=self,
                                           reversible=self.is_reversible, dt_traj=self.dt_traj)
-        estimator.estimate(dtrajs)
+        estimator.estimate(self.discrete_trajectories_full)
         return estimator.model
 
     def coarse_grain(self, ncoarse, method='hmm'):
@@ -828,7 +791,7 @@ class _MSMEstimator(_Estimator, _MSM):
     ################################################################################
 
     def cktest(self, nsets, memberships=None, mlags=10, conf=0.95, err_est=False,
-               n_jobs=1, show_progress=True, dtrajs=None):
+               n_jobs=1, show_progress=True):
         """ Conducts a Chapman-Kolmogorow test.
 
         Parameters
@@ -853,9 +816,7 @@ class _MSMEstimator(_Estimator, _MSM):
             how many jobs to use during calculation
         show_progress : bool, optional
             Show progress bars for calculation?
-        dtrajs : list containing ndarrays(dtype=int) or ndarray(n, dtype=int)
-            discrete trajectories, stored as integer ndarrays (arbitrary size)
-            or a single ndarray for only one trajectory.
+
         Returns
         -------
         cktest : :class:`ChapmanKolmogorovValidator <pyemma.msm.ChapmanKolmogorovValidator>`
@@ -875,14 +836,13 @@ class _MSMEstimator(_Estimator, _MSM):
             134: 174105
 
         """
-        dtrajs = self._check_dtrajs_arg(dtrajs)
         from pyemma.msm.estimators import ChapmanKolmogorovValidator
         if memberships is None:
             self.pcca(nsets)
             memberships = self.metastable_memberships
         ck = ChapmanKolmogorovValidator(self, self, memberships, mlags=mlags, conf=conf,
                                         n_jobs=n_jobs, err_est=err_est, show_progress=show_progress)
-        ck.estimate(dtrajs)
+        ck.estimate(self._dtrajs_full)
         return ck
 
 
@@ -894,7 +854,8 @@ class MaximumLikelihoodMSM(_MSMEstimator):
     def __init__(self, lag=1, reversible=True, statdist_constraint=None,
                  count_mode='sliding', sparse=False,
                  connectivity='largest', dt_traj='1 step', maxiter=1000000,
-                 maxerr=1e-8, score_method='VAMP2', score_k=10):
+                 maxerr=1e-8, score_method='VAMP2', score_k=10,
+                 mincount_connectivity='1/n'):
         r"""Maximum likelihood estimator for MSMs given discrete trajectory statistics
 
         Parameters
@@ -996,6 +957,12 @@ class MaximumLikelihoodMSM(_MSMEstimator):
             The maximum number of eigenvalues or singular values used in the
             score. If set to None, all available eigenvalues will be used.
 
+        mincount_connectivity : float or '1/n'
+            minimum number of counts to consider a connection between two states.
+            Counts lower than that will count zero in the connectivity check and
+            may thus separate the resulting transition matrix. The default
+            evaluates to 1/nstates.
+
         References
         ----------
         .. [1] H. Wu and F. Noe: Variational approach for learning Markov processes from time series data
@@ -1004,7 +971,8 @@ class MaximumLikelihoodMSM(_MSMEstimator):
         """
         super(MaximumLikelihoodMSM, self).__init__(lag=lag, reversible=reversible, count_mode=count_mode,
                                                    sparse=sparse, connectivity=connectivity, dt_traj=dt_traj,
-                                                   score_method=score_method, score_k=score_k)
+                                                   score_method=score_method, score_k=score_k,
+                                                   mincount_connectivity=mincount_connectivity)
 
         self.statdist_constraint = _types.ensure_ndarray_or_None(statdist_constraint, ndim=None, kind='numeric')
         if self.statdist_constraint is not None:  # renormalize
@@ -1114,7 +1082,6 @@ class MaximumLikelihoodMSM(_MSMEstimator):
 
         # Done. We set our own model parameters, so this estimator is
         # equal to the estimated model.
-        # TODO: remove in a future version.
         self._dtrajs_full = dtrajs
         self._connected_sets = msmest.connected_sets(self._C_full)
         self.set_model_params(P=P, pi=statdist_active, reversible=self.reversible,
@@ -1127,7 +1094,6 @@ class MaximumLikelihoodMSM(_MSMEstimator):
 
     # TODO: change to statistically effective count matrix!
     @property
-    @deprecated('Please use compute_effective_count_matrix in the future.')
     def effective_count_matrix(self):
         """Statistically uncorrelated transition counts within the active set of states
 
@@ -1146,33 +1112,6 @@ class MaximumLikelihoodMSM(_MSMEstimator):
         return Ceff
         # return self._C_active / float(self.lag)
 
-    # TODO: change to statistically effective count matrix!
-    def compute_effective_count_matrix(self, dtrajs):
-        """Statistically uncorrelated transition counts within the active set of states
-
-        You can use this count matrix for Bayesian estimation or error perturbation.
-
-        Parameters
-        ----------
-        dtrajs : list containing ndarrays(dtype=int) or ndarray(n, dtype=int)
-            discrete trajectories, stored as integer ndarrays (arbitrary size)
-            or a single ndarray for only one trajectory.
-
-        References
-        ----------
-        [1] Noe, F. (2015) Statistical inefficiency of Markov model count matrices
-            http://publications.mi.fu-berlin.de/1699/1/autocorrelation_counts.pdf
-
-        """
-        self._check_is_estimated()
-        dtrajs = ensure_dtraj_list(dtrajs)
-        import msmtools.estimation as msmest
-        Ceff_full = msmest.effective_count_matrix(dtrajs, self.lag)
-        from pyemma.util.linalg import submatrix
-        Ceff = submatrix(Ceff_full, self.active_set)
-        return Ceff
-        # return self._C_active / float(self.l
-
 
 @fix_docs
 @aliased
@@ -1181,7 +1120,8 @@ class OOMReweightedMSM(_MSMEstimator):
 
     def __init__(self, lag=1, reversible=True, count_mode='sliding', sparse=False, connectivity='largest',
                  dt_traj='1 step', nbs=10000, rank_Ct='bootstrap_counts', tol_rank=10.0,
-                 score_method='VAMP2', score_k=10):
+                 score_method='VAMP2', score_k=10,
+                 mincount_connectivity='1/n'):
         r"""Maximum likelihood estimator for MSMs given discrete trajectory statistics
 
         Parameters
@@ -1272,6 +1212,12 @@ class OOMReweightedMSM(_MSMEstimator):
             The maximum number of eigenvalues or singular values used in the
             score. If set to None, all available eigenvalues will be used.
 
+        mincount_connectivity : float or '1/n'
+            minimum number of counts to consider a connection between two states.
+            Counts lower than that will count zero in the connectivity check and
+            may thus separate the resulting transition matrix. The default
+            evaluates to 1/nstates.
+
         References
         ----------
         .. [1] H. Wu and F. Noe: Variational approach for learning Markov processes from time series data
@@ -1287,7 +1233,8 @@ class OOMReweightedMSM(_MSMEstimator):
 
         super(OOMReweightedMSM, self).__init__(lag=lag, reversible=reversible, count_mode=count_mode, sparse=sparse,
                                                connectivity=connectivity, dt_traj=dt_traj,
-                                               score_method=score_method, score_k=score_k)
+                                               score_method=score_method, score_k=score_k,
+                                               mincount_connectivity=mincount_connectivity)
         self.nbs = nbs
         self.tol_rank = tol_rank
         self.rank_Ct = rank_Ct
@@ -1367,7 +1314,7 @@ class OOMReweightedMSM(_MSMEstimator):
 
         # Done. We set our own model parameters, so this estimator is
         # equal to the estimated model.
-        self._dtrajs_full = dtrajs  # TODO: remove this in a future version.
+        self._dtrajs_full = dtrajs
         from pyemma.util.discrete_trajectories import index_states
         self._active_state_indexes = index_states(dtrajs, subset=self.active_set)
         self._connected_sets = msmest.connected_sets(self._C_full)
@@ -1456,7 +1403,8 @@ class AugmentedMarkovModel(MaximumLikelihoodMSM):
 
     def __init__(self,  lag=1, count_mode='sliding', connectivity='largest',
                  dt_traj='1 step',
-                 E=None, m=None, w=None, eps=0.05, support_ci=1.00, maxiter=500, debug=False, max_cache = 3000):
+                 E=None, m=None, w=None, eps=0.05, support_ci=1.00, maxiter=500, debug=False, max_cache=3000,
+                 mincount_connectivity='1/n'):
         r"""Maximum likelihood estimator for AMMs given discrete trajectory statistics and expectation values from experiments
 
         Parameters
@@ -1539,6 +1487,13 @@ class AugmentedMarkovModel(MaximumLikelihoodMSM):
         max_cache : int, default=3000
           Maximum size (in megabytes) of cache when computing R tensor (Supporting information in [1]).
 
+        mincount_connectivity : float or '1/n'
+            minimum number of counts to consider a connection between two states.
+            Counts lower than that will count zero in the connectivity check and
+            may thus separate the resulting transition matrix. The default
+            evaluates to 1/nstates.
+
+
         References
         ----------
         .. [1] Olsson, Wu, Paul, Clementi and Noe "Combining Experimental and Simulation Data via Augmented Markov Models"
@@ -1590,6 +1545,7 @@ class AugmentedMarkovModel(MaximumLikelihoodMSM):
         self.debug = debug
         self._max_cache = max_cache
         self._is_estimated = False
+        self.mincount_connectivity = mincount_connectivity
 
     def _log_likelihood_biased(self, C, T, E, mhat, ws):
         """ Evaluate AMM likelihood. """