+++ pyemma/msm/tests/test_cktest.py
@@ -34,7 +34,7 @@ from msmtools.estimation import count_matrix, largest_connected_set, largest_con
 from pyemma.util.numeric import assert_allclose
 from pyemma.msm.tests.birth_death_chain import BirthDeathChain
 from pyemma.msm import estimate_markov_model
-from six.moves import range
+
 
 import sys
 on_win = sys.platform == 'win32'
@@ -148,7 +148,6 @@ class TestCK_MSM(unittest.TestCase):
         p_MD = np.vstack([ck.estimates[:, 0, 0], ck.estimates[:, 2, 2]]).T
         assert_allclose(p_MD, self.p_MD)
 
-    @unittest.skipIf(on_win, "known to fail for pytest issue")
     def test_cktest_njobs_2(self):
         # introduce a (fake) third set in order to model incomplete partition.
         memberships = np.array([[1, 0, 0],
@@ -164,7 +163,6 @@ class TestCK_MSM(unittest.TestCase):
         p_MD = np.vstack([ck.estimates[:, 0, 0], ck.estimates[:, 2, 2]]).T
         assert_allclose(p_MD, self.p_MD)
 
-    @unittest.skipIf(on_win, "known to fail for pytest issue")
     def test_cktest_njobs_3(self):
         # introduce a (fake) third set in order to model incomplete partition.
         memberships = np.array([[1, 0, 0],