+++ pyemma/util/tests/statistics_test.py
@@ -27,7 +27,7 @@ from __future__ import absolute_import
 import unittest
 from pyemma.util import statistics
 import numpy as np
-from six.moves import range
+
 
 
 class TestStatistics(unittest.TestCase):